requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.13
aiohttp-jinja2==1.6
aiosignal==1.3.2
attrs==25.3.0
binance-futures-connector==4.1.0
certifi==2025.6.15
charset-normalizer==3.4.2
click==8.2.1
frozenlist==1.7.0
idna==3.10
Jinja2==3.1.6
linkify-it-py==2.0.3
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdit-py-plugins==0.4.2
mdurl==0.1.2
msgpack==1.1.1
multidict==6.4.4
platformdirs==4.3.8
propcache==0.3.2
pycryptodome==3.23.0
Pygments==2.19.1
requests==2.32.4
rich==14.0.0
textual==3.4.0
textual-dev==1.7.0
textual-serve==1.1.2
typing_extensions==4.14.0
uc-micro-py==1.0.3
urllib3==2.4.0
websocket-client==1.8.0
yarl==1.20.1
